Added support for a new buffer that executes trajectories with frame durations rather than frame speeds.
This change is a step toward getting proper velocity control up-and-running in our trajectory parser. Whereas before we would have had to use the setSpeed() method to control our trajectory timing, a difficult feat because it only allowed control of the END-EFFECTOR speed (not joint speeds, which are much simpler) of the arm, the new buffer will allow the user to control the time duration of each frame of movement.