Added safety code to joint buffers, tested in RobotStudio.
There is a more stringent need for safety now that we're controlling by timestep; in particular, the same miniscule timestep can result in either a dangerous fast motion or a gentle slow motion depending on the distance traveled.

In addition to some corner-case checking code which should have been there (eg. what happens when execute is called on an empty buffer), added code so that on executing the time buffer, then robot will first move to assume the first frame of the animation using whatever currentSpeed is set to, only THEREAFTER using the times in bufferJointTimes{i} to control the movement time.